# Click-Reserva/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClickReserva API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reservations:
    get:
      operationId: getReservations
      tags: [reservations]
      summary: List reservations (optionally filtered)
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: roomId
          in: query
          schema:
            type: integer
        - name: professorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a new reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationBody"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "409":
          description: Conflict or blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reservations/{reservationId}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get a single reservation
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    delete:
      operationId: cancelReservation
      tags: [reservations]
      summary: Cancel a reservation
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /reservations/{reservationId}/approve:
    post:
      operationId: approveReservation
      tags: [reservations]
      summary: Approve a pending reservation (coordinator only)
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reservations/{reservationId}/reject:
    post:
      operationId: rejectReservation
      tags: [reservations]
      summary: Reject a pending reservation (coordinator only)
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reservations/{reservationId}/confirm-presence:
    post:
      operationId: confirmPresence
      tags: [attendance]
      summary: Confirm presence for a reservation
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Presence confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Cannot confirm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Reservation:
      type: object
      properties:
        id:
          type: integer
        professorId:
          type: integer
        professorName:
          type: string
        roomId:
          type: integer
        roomNumber:
          type: string
        roomName:
          type: string
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        subject:
          type: string
        classGroup:
          type: string
        status:
          type: string
          enum: [pending, confirmed, realized, cancelled, no_show, rejected]
        confirmedPresence:
          type: boolean
        confirmedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - professorId
        - professorName
        - roomId
        - roomNumber
        - roomName
        - date
        - startTime
        - endTime
        - subject
        - classGroup
        - status
        - confirmedPresence
        - createdAt
    CreateReservationBody:
      type: object
      properties:
        roomId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        subject:
          type: string
        classGroup:
          type: string
        professorId:
          type: integer
          description: Optional - coordinator can create on behalf of a professor
      required:
        - roomId
        - date
        - startTime
        - endTime
        - subject
        - classGroup
...